Refactor API routes and improve code reuse
Use `postFetch` and `getFetch` functions for API calls in multiple
files, enhancing code readability and reducing duplication. This change
replaces the direct usage of `fetch` with these custom functions in the
authentication, gallery, and profile API routes. It also includes the
addition of the `headerCookies` parameter in the `postFetch` call to
pass the request headers for accessing the login API. The introduction
of these reusable functions simplifies the codebase and promotes
consistent API handling across the application.

Resolves: #123